Update code 15/11

+++ WebCalenderAPI/Controllers/SchedulesController.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
+using WebCalenderAPI.Helper;
 using WebCalenderAPI.Models;
 using WebCalenderAPI.Services;
 
@@ -8,19 +9,27 @@ namespace WebCalenderAPI.Controllers
 {
     [Route("api/[controller]")]
     [ApiController]
-    //[Authorize]
+    [Authorize]
     public class SchedulesController : ControllerBase
     {
         private readonly IScheduleRepository _scheduleRepository;
+        private readonly ICacheService _cacheService;
+        private readonly TokenHelper _tokenHelper;
 
-        public SchedulesController(IScheduleRepository scheduleRepository)
+        public SchedulesController(IScheduleRepository scheduleRepository, ICacheService cacheService, TokenHelper tokenHelper)
         {
             _scheduleRepository = scheduleRepository;
+            _cacheService = cacheService;   
+            _tokenHelper = tokenHelper;
         }
         [HttpGet]
-        public IActionResult GetAll()
+        public async Task<IActionResult> GetAll()
         {
             try {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null) {
+                    return tokenResult;
+                }
                 return Ok(_scheduleRepository.GetALl());
             } catch
             {
@@ -28,26 +37,67 @@ namespace WebCalenderAPI.Controllers
             }
         }
 
+        private async Task<IActionResult> HandleTokenRefresh()
+        {
+            string accessToken = _cacheService.GetData<String>("accessToken");
+
+            string refreshToken = _cacheService.GetData<String>("refreshToken");
+
+            var tokenModel = new TokenModel
+            {
+                AccessToken = accessToken,
+                RefreshToken = refreshToken
+            };
+            var checkToken = await _tokenHelper.RenewToken(tokenModel);
+            if (checkToken.Status == "401")
+            {
+                _cacheService.RemoveData("accessToken");
+                _cacheService.RemoveData("refreshToken");
+                return Unauthorized(new CheckTokenResult
+                {
+                    Status = "401",
+                    Error = checkToken.Error
+                });
+            }
+            else if (checkToken.Status == "200")
+            {
+                var acccessToken = checkToken.AccessToken;
+                _cacheService.SetData("accessToken", acccessToken);
+            }
+            return null;
+        }
+
         [HttpGet("getSchedule/{userId}")]
-        public IActionResult GetScheduleByUserId(int userId)
+        public async Task<IActionResult> GetScheduleByUserId(int userId)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.getAllScheduleByUser(userId));
             }
             catch
             {
                 return StatusCode(StatusCodes.Status500InternalServerError);
             }
-
         }
 
 
         [HttpGet("{id}")]
-        public IActionResult GetById(int id)
+        public async Task<IActionResult> GetById(int id)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.GetById(id));
             }
             catch
@@ -57,10 +107,16 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpGet("dateTime/{dateTime}")]
-        public IActionResult GetByDateTime(DateTime dateTime)
+        public async Task<IActionResult> GetByDateTime(DateTime dateTime)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.GetByDate(dateTime));
             }
             catch
@@ -69,9 +125,14 @@ namespace WebCalenderAPI.Controllers
             }
         }
         [HttpGet("getAllDate/{dateTime}")]
-        public IActionResult GetByDateTimeWithHavingReason(DateTime dateTime) {
+        public async Task<IActionResult> GetByDateTimeWithHavingReason(DateTime dateTime) {
             try {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
 
+                }
                 return Ok(_scheduleRepository.getByDateWithReason(dateTime));
             }
             catch
@@ -82,10 +143,16 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpGet("getAllDateByUserId")]
-        public IActionResult GetByDateTimeByUserIdWithHavingReason([FromQuery]DateTime dateTime,[FromQuery] int userId)
+        public async Task<IActionResult> GetByDateTimeByUserIdWithHavingReason([FromQuery]DateTime dateTime,[FromQuery] int userId)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.getByDateWithReasonWithUserId(userId, dateTime));
             }
             catch
@@ -96,10 +163,16 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpGet("getAllDateByUserIdWithoutReason")]
-        public IActionResult GetByDateTimeByUserIdWithoutReason([FromQuery] DateTime dateTime, [FromQuery] int userId)
+        public async Task<IActionResult> GetByDateTimeByUserIdWithoutReason([FromQuery] DateTime dateTime, [FromQuery] int userId)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.getAllScheduleWithoutReason(userId, dateTime));
             }
             catch
@@ -110,8 +183,14 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpPut("{id}")]
-        public IActionResult Update(int id,ScheduleVM scheduleVM)
+        public async Task<IActionResult> Update(int id,ScheduleVM scheduleVM)
         {
+            var tokenResult = await HandleTokenRefresh();
+            if (tokenResult != null)
+            {
+                return tokenResult;
+
+            }
             if (id != scheduleVM.id)
             {
                 return BadRequest();
@@ -128,10 +207,16 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpDelete("{id}")]
-        public IActionResult Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 _scheduleRepository.Delete(id);
                 return Ok();
             }
@@ -143,10 +228,16 @@ namespace WebCalenderAPI.Controllers
 
         [HttpPost]
         
-        public IActionResult Add(ScheduleAdd scheduleAdd)
+        public async Task<IActionResult> Add(ScheduleAdd scheduleAdd)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.Add(scheduleAdd));
             }
             catch
@@ -156,10 +247,16 @@ namespace WebCalenderAPI.Controllers
         }
 
         [HttpPost("insertByDate")]
-        public IActionResult AddScheduleWithDate([FromBody] ScheduleMeta meata)
+        public async Task<IActionResult> AddScheduleWithDate([FromBody] ScheduleMeta meata)
         {
             try
             {
+                var tokenResult = await HandleTokenRefresh();
+                if (tokenResult != null)
+                {
+                    return tokenResult;
+
+                }
                 return Ok(_scheduleRepository.AddScheduleWithDate(meata));
             }
             catch
